pyproject.toml
[project]
name = "agentic-log-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "instructor>=1.9.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
